refactor(results): table as a separate component
