+++ components/ResultsPage/OptionSelectFilter.tsx
@@ -1,0 +1,8 @@
+import { FilterProps } from "./resultsTypes"
+
+
+const OptionSelectFilter = ({category_name,values} : FilterProps) => {
+    return(<span>{category_name}</span>)
+}
+
+export default OptionSelectFilter
