+++ components/ResultsPage/ResultTable.module.scss
@@ -1,0 +1,14 @@
+@use '../../styles/shared.scss' as shared;
+
+.results {
+    grid-column: 1/-1;
+    font-size: shared.$desktop-reading-font-size
+}
+.result_table {
+    width: 100%;
+    border-collapse: collapse;
+    tr { 
+    border: solid;
+    border-width: 1px 0;
+    }
+}
